The path to the AIM directory is added by dynare_config.m
for start, options_.useAIM = 0 is set by default in global_initialization.m so that system uses mjdgges by default. 

If AIM is to be used, options_.useAIM = 1 needs to be set either in the model <>.mod file, before invoking, estimate and/or stoch_simul, or by issuing appropriate command for estimate and/or stoch_simul. 
GP July '08

git-svn-id: https://www.dynare.org/svn/dynare/dynare_v4@1966 ac1d8469-bf42-47a9-8791-bf33cf982152