Added a mex  file for computing Sobol quasi  monte carlo sequences (in an hypercube of max dimension 1111).
Because of a fortran 77 source  compiled with gfortran (g77 is no more
available on recent  debian system), we have to  override the gfortran
library shipped with matlab. This can be done using the LD_PRELOAD
environment  variable  in  the  .bashrc *or*  removing  the  libraries
distributed with  matlab *or*  modifying the script  .matlab7rc.sh. We
may provide  a (python or whatever)  script to do this  when dynare is
installed.


git-svn-id: https://www.dynare.org/svn/dynare/trunk@2574 ac1d8469-bf42-47a9-8791-bf33cf982152