Fixed bug in dynare_resolve (wrong calling sequence introduced in commit #013c599ec92f7d6e5fc3f351a58d9aa5ba401410).
Removed globals from DsgeVarLikelihood and changed the calling sequence. As in DsgeLikelihood, the penalty is now a
persistent variable.

Added a global structure for the data: dataset_.

Removed globals from dsgevar_posterior_density and mode_check.

Simplification of the clode, definition of the variable objective_function at the top of dynare_estimation_1 (equal
to 'DsgeLikelihood' or 'DsgeVarLikelihood').