Changed subsref overloaded method in dynDate class.
Allow to populate an empty dynDate object or update a dynDate object. For instance, if we instantiate
the dynDate class as follows:

a = dynDate()

a is an empty dynDate object. We can "populate" this object as follows

a('1938Q4')

Note however that a ins not modified in place (due to a limitation of Octave's class design), a copy
has to be made.