Add function for displaying estimation results
The commit moves the creation of results tables, both the ones displayed
and printed to LaTeX, from dynare_estimation_1.m to a separate function.
This factorization later allows to use this function to display results
from other estimation routines. At the same time, it removes the display
of t-statistics for posterior estimates ( #339 )