Changed behaviour of dynSeries objects.
Let ts be a dynSeries object. The following syntaxes are equivalent:

a = ts.lead
a = ts(1)

b = ts.lag(2)
b = ts(-2)

Advantage: If (some of) the variables used in the model block are known as dynSeries in matlab's workspace, then we can create new dynSeries objects with simple copy/pastes of the model's equations (because dynSeries objects understands leads and lags as Dynare's preprocessor).